Crate bounce

The uncomplicated Yew State management library.

Re-exports

pub use states::atom::Atom;

pub use states::slice::Slice;

Modules

prelude

Default Bounce exports.

Structs

Artifact

A component to register an artifact.

ArtifactProps

Properties of the Artifact Component.

BounceRoot

A <BounceRoot />.

BounceRootProps

Properties for BounceRoot.

BounceStates

A type to access states under a bounce root.

UseAtomHandle

A handle returned by use_atom.

UseSliceHandle

A handle returned by use_slice.

Enums

Deferred

A deferred result type for future notions.

Traits

CloneAtom

A trait to provide cloning on atoms.

CloneSlice

A trait to provide cloning on slices.

FutureNotion

A trait to implement a Future-backed notion.

InputSelector

An auto-updating derived state, similar to Selector, but with an input.

Observed

A trait to be notified when the state value changes.

Selector

An auto-updating derived state.

WithNotion

A trait to apply a notion on a state.

Functions

use_artifacts

A hook to read all artifacts of the current artifact type.

use_atom

A hook to connect to an Atom.

use_atom_setter

A hook to produce a setter function for an Atom.

use_atom_value

A read-only hook to connect to the value of an Atom.

use_future_notion_runner

A hook to create a function that when called, runs a FutureNotion with provided input.

use_input_selector_value

A hook to connect to an InputSelector.

use_notion_applier

A hook to create a function that applies a Notion.

use_selector_value

A hook to connect to a Selector.

use_slice

A hook to connect to a Slice.

use_slice_dispatch

A hook to produce a dispatch function for a Slice.

use_slice_value

A read-only hook to connect to the value of a Slice.

Attribute Macros

future_notion

A future-based notion that notifies states when it begins and finishes.

Derive Macros

Atom

A simple state that is Copy-on-Write and notifies registered hooks when prev_value != next_value.

Slice

A reducer-based state that is Copy-on-Write and notifies registered hooks when prev_value != next_value.